enhanced

+++ app/src/main/java/com/example/mybabyvaxadmin/models/Dose.kt
@@ -3,7 +3,7 @@ package com.example.mybabyvaxadmin.models
 data class Dose(
     var id: String? = null,
     var name: String? = null,
-    var intervalNumber: Int? = null,
+    var intervalNumber: Double? = null,
     var intervalUnit: String? = null,
     var description: String? = null
 )
